backend/requirements.txt
# NEXUS PROTOCOL - Backend Dependencies
# FastAPI + Server
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
python-multipart==0.0.6

# Kafka (Confluent)
confluent-kafka==2.3.0

# Datadog Observability
datadog==0.48.0
ddtrace==2.5.0

# Google Cloud / Vertex AI (Gemini)
google-cloud-aiplatform==1.38.0
google-auth==2.25.0

# ElevenLabs Voice
elevenlabs==1.0.0

# HTTP Client
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.0
pytest-asyncio==0.21.0
